Fix for differing window sizes on staking form
Bug Explanation:
When changing window size the 'MAX AMOUNT' button does not
remain in the desired location.

Fix:
Changed the column widths on small and medium devices to use
justifyContent = space-between instead of padding. Changed the
grid item size for the button to a fixed value equal to the size of
the box.