Added feature for helper text in staking form
Problem:
When user clicks the 'MAX AMOUNT' button with no connected
wallet, a helper text of 'Enter a valid token amount' appears. The
actual error is there is no account to pull the amount from.

Fix:
Changed the conditions of the helper text to include that
in the case of the users token balance being null, the message
'Please connect with yoroi or nautilus to use MaxAmount' is
displayed.