Added button to staking form to cancel
Problem: The current way of exiting the form is by clicking outside
the modal. This is not intuitive and led to me feeling like there was
no way out of the form.

Fix: Added a button to allow the user to close the form. Also updated themes to include a quinary red cancel palate.